Added unit test case for the reinforcement python program
